optimizer: address code-review findings on the ValueGraph const-folding
Cleanups surfaced reviewing the mut_escaped / HFS-forward / const-folding
work; all behavior-preserving (golden fixtures byte-identical, full e2e
green):

- Extract niri's pure constant-evaluation primitives (`Value`,
  `eval_binary` / `eval_unary` / `prim_of` / `value_to_arena_kind` and the
  int/float/bool/char helpers) into a new lower `const_eval` module, so the
  ValueGraph builder folds arithmetic without depending up on `niri`.
  `niri` re-exports `Value` for the public API.
- Apply niri's `is_int_prim` / float-prim consistency filter in the one
  shared `value_kind_to_const` bridge (used by the builder's `value_to_const`
  and `store_load_forward`'s `synth_literal_kind`), so the value-graph folder
  folds exactly the set niri's CTFE folds — no wide-int (`i128`/`u128`)
  divergence.
- Look through `Newtype` in `reference_pointee_struct_key` so same-pointee
  references whose pointee is a newtype still share an alias group (closes a
  latent `mut_escaped` group-closure gap).
- Reuse the alias groups `build_alias_info` already computes in
  `build_mut_escaped` instead of recomputing `same_pointee_reference_edges`
  + `collect_alias_groups`; simplify the `provably_immutable` predicate.
- Share `first_param_types` and a `projection_root_local` helper between
  `alias` and `copy_prop`; extract a shared `method_mutates_receiver`
  (parameterized on the unknown-callee verdict the two passes intentionally
  differ on).
- Borrow the struct-field map in `CallImmutability::walk` instead of cloning
  per cache miss.
